Ensure the longest possible alias match is found
A previous fix to the regex used to match aliases changed the regex to
escape any non-alphanumeric characters to avoid issues with commands
containing regex special characters. This meant that any dashes/spaces
in the command were ignored, and only the first word in the command
could ever be matched. This commit modifies the escape sequence to only
escape those characters which would explicitly cause issues if they are
not escaped.

Fix #2